# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Item API
  version: 1.0.0
paths:
  /api/items:
    get:
      operationId: listItems
      responses:
        '200':
          description: Item list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
    post:
      operationId: createItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemRequest'
      responses:
        '201':
          description: Item created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          description: Validation error
  /api/items/{id}:
    get:
      operationId: getItem
      parameters:
        - $ref: '#/components/parameters/ItemId'
      responses:
        '200':
          description: Item found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          description: Item not found
    put:
      operationId: updateItem
      parameters:
        - $ref: '#/components/parameters/ItemId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemRequest'
      responses:
        '200':
          description: Item updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          description: Validation error
        '404':
          description: Item not found
    delete:
      operationId: deleteItem
      parameters:
        - $ref: '#/components/parameters/ItemId'
      responses:
        '204':
          description: Item deleted
        '404':
          description: Item not found
components:
  parameters:
    ItemId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
  schemas:
    Item:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    ItemRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1